QRScanner/requirements.txt
kivy>=2.2.0
opencv-python>=4.8.0
pyzbar>=0.1.9
Pillow>=10.0.0
numpy>=1.24.0
pyjnius>=1.5.0; platform_system=='Android'
